rename std folder to stdtypes
