+++ SofaKernel/modules/SofaTypes/src/sofa/types/stdtypes/fixed_array_algorithms.h
@@ -23,7 +23,7 @@
 
 #include <sofa/types/config.h>
 
-#include <sofa/types/std/fixed_array.h>
+#include <sofa/types/stdtypes/fixed_array.h>
 
 namespace sofa::types::stdtypes::pairwise
 {
